app/requirements/base.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
cloudinary==1.26.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.2.3
django-cors-headers==3.8.0
django-crontab==0.7.1
django-filter==22.1
django-mirage-field==1.2.2
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
numpy==1.24.2
openapi==1.1.0
packaging==21.0
pandas==1.5.3
Pillow==8.3.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.11.0
PyJWT==2.1.0
pyotp==2.6.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
tqdm==4.62.3
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
zipp==3.6.0
